+++ src/main/java/recruitment/controller/ApplicationController.java
@@ -6,6 +6,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import recruitment.domain.*;
+import recruitment.dto.ApplicationDetailedDto;
+import recruitment.dto.ApplicationDto;
+import recruitment.domain.Job;
 import recruitment.exception.ResourceAlreadyExist;
 import recruitment.exception.ResourceNotFound;
 import recruitment.repository.ApplicationRepository;
