+++ src/main/java/recruitment/controller/JobController.java
@@ -6,8 +6,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import recruitment.domain.Company;
 import recruitment.domain.Job;
-import recruitment.domain.JobDto;
-import recruitment.domain.JobSimpleDto;
+import recruitment.dto.JobDto;
+import recruitment.dto.JobSimpleDto;
 import recruitment.exception.ResourceNotFound;
 import recruitment.repository.CompanyRepository;
 import recruitment.repository.JobRepository;
