+++ src/test/java/recruitment/controller/ApplicationControllerTest.java
@@ -14,6 +14,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
 import recruitment.domain.*;
+import recruitment.dto.ApplicationDetailedDto;
+import recruitment.dto.ApplicationDto;
+import recruitment.dto.JobSimpleDto;
 import recruitment.exception.ResourceNotFound;
 import recruitment.repository.ApplicationRepository;
